Adding the first web bits.
Note that this is just an import of the block checker, but it does put the bits in place for doing the twitter oauth stuff itself in place. Need to change the protocol for talking to the unix socket, change UI, and sanitize/change how we talk to the database that stores tokens.

Eventually the tokens are going to be stored outside of MySQL, but this code is here for now, and better done than perfect. ;)