Clean index.html a bit
I'm pretty sure git will believe it's a rewrite.
It's html. Except when told otherwise, blank space are collapsed into a single space. So I inserted plenty of new line in the text so it should fit more or less in 80 columns. 

Indent should be 4 space all the way. I know, it's not conform to style(9). I used that because it's the default in github online editor.
Most of the work was to remove the font, b and center tag and create a warning class. Now, everything look clean :)